# openapi3.yaml
openapi: 3.0.1
info:
  title: dem-heights
  description: map-colonies service for extracting elevations from DTM/DSM geotiffs
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

paths:
  /heights/points:
    post:
      operationId: getPoints
      tags:
        - heights
      summary: get points heights
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
  /heights/path:
    post:
      operationId: getPath
      tags:
        - heights
      summary: get path heights
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
  /heights/polygon:
    post:
      operationId: getPolygon
      tags:
        - heights
      summary: get polygon heights
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
  /heights:
    post:
      operationId: getHeights
      tags:
        - heights
      summary: get heights for any given geojson
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
  /heights/{longitude}/{latitude}:
    get:
      operationId: getHeight
      tags:
        - heights
      summary: get height at a specific (lon,lat)
      parameters:
        - name: longitude
          in: path
          required: true
          schema:
            type: number
          example: 35.034
        - name: latitude
          in: path
          required: true
          schema:
            type: number
          example: 32.691
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/height'
components:
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    height:
      type: object
      required:
        - dem
      properties:
        dem:
          type: number
          format: int64